Move project model to its own file
